requirements.txt
Django==3.1.4
django-mqueue==0.12.0
django-instant==2.1.2